stash: porting to windows

+++ rig/rut/rut-bitmask.h
@@ -63,12 +63,12 @@ C_BEGIN_DECLS
 typedef struct _rut_bitmask_imaginary_type_t *RutBitmask;
 
 /* These are internal helper macros */
-#define _rut_bitmask_to_number(bitmask) ((unsigned long)(*bitmask))
+#define _rut_bitmask_to_number(bitmask) ((unsigned long)(uintptr_t)(*bitmask))
 #define _rut_bitmask_to_bits(bitmask) (_rut_bitmask_to_number(bitmask) >> 1UL)
 /* The least significant bit is set to mark that no array has been
    allocated yet */
 #define _rut_bitmask_from_bits(bits)                                           \
-    ((RutBitmask)((((unsigned long)(bits)) << 1UL) | 1UL))
+    ((RutBitmask)(uintptr_t)((((unsigned long)(bits)) << 1UL) | 1UL))
 
 /* Internal helper macro to determine whether this bitmask has a
    c_array_t allocated or whether the pointer is just used directly */
